Add files via upload
This is the dataset used to generate Table 9.1 in the paper "An MBO scheme for clustering and semi-supervised clustering of signed networks" by Mihai Cucuringu, Andrea Pizzoferrato, and Yves van Gennip. The data, which are publicly available, were originally downloaded from https://www.census.gov/population/www.cen2000/ctytoctyflow/index.html but the link appears to be broken (verified on 10/08/2020). The new link for the data is https://www.census.gov/data/tables/2000/demo/geographic-mobility/county-to-county-migration-flows.html but the formatting is different. For this reason, we have decided to include on GitHub the Matlab files we produced so that the interested reader can have access to the original dataset directly in matrix form.